move a few tests from tests/ to src/ so it's all in the same place and consistent
side benefit it all looks better in vscode. i'm sure there's something else i could have fixed but meh
